Added specific JSON Decode error exception
This happens when the tfp claim is empty or invalid and it might result in an incorrect URL for openid-configuration. This would the return a 404 and if parsed bubbles up an error saying `Expecting value: line 1 column 1 (char 0)` on the `json()` call on the response.
Catching the JSONDecode error is more specific and highlights potential underlying root cause better.